+++ Sources/DistDeployer/QueueServerConfigurationLocation.swift
@@ -1,0 +1,8 @@
+import Foundation
+import TypedResourceLocation
+
+public typealias QueueServerConfigurationLocation = TypedResourceLocation<QueueServerConfigurationLocationType>
+
+public final class QueueServerConfigurationLocationType: ResourceLocationType {
+    public static let name = "queue server configuration"
+}